moved InstallFont to the install folder
